feat: change chakra component to tailwind
