+++ package.json
@@ -20,12 +20,14 @@
     "@emotion/react": "11.10.6",
     "@emotion/styled": "11.10.6",
     "@formatjs/intl-localematcher": "^0.2.32",
+    "@hello-pangea/dnd": "^16.2.0",
     "@tanstack/react-query": "^4.27.0",
     "@types/node": "18.14.6",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
     "@types/react-slick": "^0.23.10",
     "accept-language": "^3.0.18",
+    "add": "^2.0.6",
     "axios": "^1.3.4",
     "date-fns": "^2.29.3",
     "dayzed": "^3.2.3",
@@ -47,6 +49,7 @@
     "prettier": "2.8.4",
     "react": "18.2.0",
     "react-beautiful-dnd": "^13.1.1",
+    "react-date-picker": "^9.2.0",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
     "react-i18next": "^12.2.0",
@@ -55,7 +58,8 @@
     "recoil": "0.7.7",
     "server-only": "^0.0.1",
     "slick-carousel": "^1.8.1",
-    "typescript": "4.9.5"
+    "typescript": "4.9.5",
+    "yarn": "^1.22.19"
   },
   "devDependencies": {
     "@babel/core": "^7.21.0",
@@ -74,7 +78,7 @@
     "@storybook/testing-library": "^0.0.13",
     "@types/i18next": "^13.0.0",
     "@types/negotiator": "^0.6.1",
-    "@types/react-beautiful-dnd": "^13.1.3",
+    "@types/react-beautiful-dnd": "^13.1.4",
     "@yarnpkg/sdks": "^3.0.0-rc.40",
     "autoprefixer": "^10.4.14",
     "babel-loader": "^8.3.0",
