+++ src/app/[lng]/vote/create/page.tsx
@@ -1,4 +1,4 @@
-import CreateVote from "@/components/Vote/CreateVote";
+import CreateVote from "@/components/vote/create/CreateVote";
 const CreateVotePage = () => {
   return (
     <>
