+++ src/common/uploadFirebase.ts
@@ -10,7 +10,6 @@ const uploadFirebase = async (file: File): Promise<string> => {
       (snapshot) => {
         const progress =
           (snapshot.bytesTransferred / snapshot.totalBytes) * 100;
-        console.log("Upload is " + progress + "% done");
       },
       (error) => {
         reject(""); // added this line
